2026-03-16-sf-bulk-api-data-loader/requirements.txt
simple-salesforce==1.12.6
python-dotenv==1.0.1
